+++ src/routes/me/index.ts
@@ -4,6 +4,7 @@ import blocked from "./blocked";
 import avatar from "./avatar";
 import { FastifyInstance, FastifyPluginOptions } from "fastify";
 import votes from "./votes";
+import starred from "./starred";
 
 export default function (
     fastify: FastifyInstance,
@@ -12,8 +13,9 @@ export default function (
 ) {
     fastify.register(status);
     fastify.register(rename);
-    fastify.register(blocked);
     fastify.register(avatar);
     fastify.register(votes);
+    fastify.register(blocked);
+    fastify.register(starred);
     done();
 }